O que precisamos é do Santander, não mais do Banespa.
